requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.1
requests>=2.31.0
trafilatura>=1.6.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
gradio>=4.0.0
